refactor(config): move space-separated deserializer to saluki-config
Extracts the space_separated_or_seq serde deserializer that previously
lived in proxy.rs into saluki_config::space_separated, where it sits
alongside duration_string as a reusable config deserialization utility.
Adds an Option variant used by ObfuscationFlatConfig in apm.rs.
